script/requirements.txt
PyYAML==5.3.1
JSON-log-formatter==0.3.0